Move metadata classes to dedicated namespace
