+++ EndlessClient/Rendering/Effects/EffectSpriteManager.cs
@@ -1,4 +1,6 @@
 ﻿using AutomaticTypeMapper;
+using EndlessClient.Rendering.Metadata;
+using EndlessClient.Rendering.Metadata.Models;
 using EOLib.Graphics;
 using Microsoft.Xna.Framework.Graphics;
 using System.Collections.Generic;
